requirements.txt
hydra-core==1.3.2
omegaconf==2.3.0
pandas==2.3.3
einops==0.8.1
matplotlib==3.10.8
tqdm==4.67.1
opencv-python==4.12.0.88
transformers==4.57.6
sentencepiece==0.2.1
networkx==3.6.1
wandb==0.24.0
scipy==1.17.0
smplx==0.1.28
pytorch-lightning==2.4.0
lmdb==1.7.5
spacy==3.8.11
tensorboard==2.20.0
tensorboardX==2.6.4
torchmetrics==1.8.2
sentence-transformers==5.1.2
peft==0.17.1
rich==14.3.3
imageio==2.37.2

git+https://github.com/otaheri/chamfer_distance
git+https://github.com/otaheri/bps_torch
git+https://github.com/openai/CLIP.git
